// cgocore-ui/src/app/features/models/loading/views/full/DatedFlightUldFullView.ts
import { DatedFlightUldForm } from 'src/app/features/models/loading/forms/DatedFlightUldForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class DatedFlightUldFullView {

id: number;
form: DatedFlightUldForm;
canUpdate: boolean;
canDelete: boolean;

/* Specific Code Start */
/* Specific Code End */
}